requirements.txt
channels==3.0.4
Django==4.0.2
django-cors-headers==3.11.0
djangorestframework==3.13.1

